# api/openapi.yaml
openapi: 3.1.0
info:
  title: Synthesis server
  version: "1.0.0"
  summary: The application service and the relay boundary of the Synthesis server.
  description: |
    The REST surface of `specifications/server/SAS-server-application-service.md`,
    the relay introspection routes of
    `specifications/server/SRB-server-relay-boundary.md`, the project-discovery
    route of `specifications/server/RSN-remote-session.md`, and the three
    WebSocket upgrades of `specifications/server/WSK-websocket.md`.

    Every route and every WebSocket upgrade but `GET /v1/health` needs the header
    `Authorization: Bearer <token>`, carrying `SYNTHESIS_SERVER_TOKEN`. On the
    application routes the token authenticates the persisted administrator user
    (SAS-FR-LFCA, SAS-FR-XRPD), which bypasses every access-control check. On the
    relay it authenticates transport access alone (RSN-FR-TWQJ): the handle
    check, the worker-ownership check, and the end-to-end client authentication
    stand beside it.

    The permission each project-scoped operation needs is named in the
    description of that operation. The relay evaluates no permission.

    The frames the three WebSocket endpoints carry are not described here: a
    WebSocket protocol is outside what OpenAPI describes, and
    `specifications/server/WSK-websocket.md` holds the frame tables.
servers:
  - url: http://localhost:8080
    description: The service as it binds by default.
security:
  - bearerAuth: []
...
paths:
...
  /v1/projects:
    get:
      tags: [Projects]
      summary: List the projects the caller may read (project.read).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The window of records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Projects]
      summary: Create a project, with its owner grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      responses:
        '201':
          description: The record that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{project_id}/grants:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Grants]
      summary: List the grants of a project (project.manage_acl).
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: The window of records.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Grant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Grants]
      summary: Create a grant (project.manage_acl).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGrant'
      responses:
        '201':
          description: The record that was written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: Read one project.
      description: Needs project.read.
      responses:
        '200':
          description: The record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Projects]
      summary: Update one project.
      description: Needs project.update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
      responses:
        '200':
          description: The record as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Projects]
      summary: Delete one project.
      description: Needs the owner or the administrator.
      parameters:
        - $ref: '#/components/parameters/Cascade'
      responses:
        '204':
          description: The record was deleted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{project_id}/grants/{grant_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/GrantId'
    get:
      tags: [Grants]
      summary: Read one grant.
      description: Needs project.manage_acl.
      responses:
        '200':
          description: The record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Grants]
      summary: Update one grant.
      description: Needs project.manage_acl.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGrant'
      responses:
        '200':
          description: The record as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Grants]
      summary: Delete one grant.
      description: Needs project.manage_acl.
      responses:
        '204':
          description: The record was deleted.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/projects/{project_id}/transfer-ownership:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [Projects]
      summary: Move the ownership of a project.
      description: Reserved to the owner and to the administrator. The owner grant moves with the ownership, and this is the one way the owner changes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferOwnership'
      responses:
        '200':
          description: The project as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/projects/{project_id}/grants/{grant_id}/revoke:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/GrantId'
    post:
      tags: [Grants]
      summary: Revoke a grant.
      description: Needs project.manage_acl. The grant supplies nothing at once and stays readable until it is deleted. The owner grant is an owner_protected conflict.
      responses:
        '200':
          description: The grant as it now stands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Grant'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /v1/roles:
    get:
      tags: [Projects]
      summary: The role-to-permission mapping.
      description: The mapping the service publishes, for project, team, and organization roles.
      responses:
        '200':
          description: The mapping.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Roles'
        '401': { $ref: '#/components/responses/Unauthenticated' }
  /v1/projects/{project_id}/permissions:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Projects]
      summary: The effective permissions of one user on one project.
      description: The caller's own permissions need project.read; another user's need project.manage_acl.
      parameters:
        - $ref: '#/components/parameters/UserIdQuery'
      responses:
        '200':
          description: The union of the permissions the grants supply.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectivePermissions'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The static token of V1 (SAS-FR-LFCA), read from `SYNTHESIS_SERVER_TOKEN`.
        On the application routes it authenticates the persisted administrator
        user; on the relay routes and the WebSocket upgrades it authenticates
        transport access alone (SRB-FR-JEBM). The service compares it in constant
        time (SAS-FR-BZUK), and it reaches no answer, no log record, and no
        persisted record (SAS-FR-PMRB).
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      description: The page size. The default is 100 and the maximum is 500.
      schema: { type: integer, minimum: 1, maximum: 500 }
    Offset:
      name: offset
      in: query
      required: false
      description: How many records to pass over before the page starts.
      schema: { type: integer, minimum: 0 }
...
    UserIdQuery:
      name: user_id
      in: query
      required: false
      description: The user whose permissions are read. It defaults to the caller.
      schema: { type: string, format: uuid }
    Cascade:
      name: cascade
      in: query
      required: false
      description: Deletes the drafts, the conversations, and the grants of the project with it.
      schema: { type: boolean, default: false }
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    GrantId:
      name: grant_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: |
        A field is absent, empty, or too long; an identifier is not a UUID; or a
        role or a permission is unknown. The codes are invalid_field,
        invalid_role, and invalid_permission.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: The request carries no accepted credential. The code is unauthenticated.
      headers:
        WWW-Authenticate:
          schema: { type: string }
          description: The scheme the route accepts.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: The principal may read the record but holds no permission for the operation. The code is forbidden.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        No record holds the identifier, or the principal may not read the
        project, which is answered the same way so the existence of a project is
        not disclosed. The code is not_found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        The write conflicts with the state the service holds. The codes are
        duplicate_id, duplicate_membership, duplicate_grant, owner_protected,
        referenced, stale_revision, conversation_locked, and
        invitation_not_pending.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - invalid_field
                - invalid_role
                - invalid_permission
                - duplicate_id
                - duplicate_membership
                - duplicate_grant
                - owner_protected
                - referenced
                - stale_revision
                - conversation_locked
                - invitation_not_pending
                - internal
            message: { type: string }
    Permission:
      type: string
      enum:
        - project.read
        - project.update
        - project.manage_acl
        - draft.read
        - draft.create
        - draft.update
        - draft.archive
        - draft.delete
        - draft.graduate
        - conversation.read
        - conversation.create
        - conversation.append
        - conversation.update
        - conversation.resolve
        - conversation.lock
    ProjectRole:
      type: string
      enum: [owner, administrator, contributor, viewer]
    MemberRole:
      type: string
      enum: [owner, administrator, member, viewer]
    TargetType:
      type: string
      enum: [user, team, organization]
...
    Project:
      type: object
      required: [id, display_name, owner_id, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        display_name:
          type: string
          maxLength: 200
          description: Presentation only. The relay never routes by this value.
        owner_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateProject:
      type: object
      required: [display_name]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string, maxLength: 200 }
        owner_id: { type: string, format: uuid }
    UpdateProject:
      type: object
      properties:
        display_name: { type: string, maxLength: 200 }
    TransferOwnership:
      type: object
      required: [new_owner_id]
      properties:
        new_owner_id: { type: string, format: uuid }
    Grant:
      type: object
      required: [id, project_id, target_type, target_id, role, permissions, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        target_type: { $ref: '#/components/schemas/TargetType' }
        target_id: { type: string, format: uuid }
        role: { $ref: '#/components/schemas/ProjectRole' }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/Permission' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
        revoked_by: { type: [string, 'null'], format: uuid }
    CreateGrant:
      type: object
      required: [target_type, target_id, role]
      properties:
        id: { type: string, format: uuid }
        target_type: { $ref: '#/components/schemas/TargetType' }
        target_id: { type: string, format: uuid }
        role:
          type: string
          enum: [administrator, contributor, viewer]
          description: The owner role is reached by a transfer of ownership alone.
    UpdateGrant:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [administrator, contributor, viewer]
...
    Roles:
      type: object
      required: [permissions, project_roles, team_roles, organization_roles]
      properties:
        permissions:
          type: array
          items: { $ref: '#/components/schemas/Permission' }
        project_roles:
          type: array
          items:
            type: object
            required: [role, permissions]
            properties:
              role: { $ref: '#/components/schemas/ProjectRole' }
              permissions:
                type: array
                items: { $ref: '#/components/schemas/Permission' }
        team_roles:
          type: array
          items: { $ref: '#/components/schemas/MemberRole' }
        organization_roles:
          type: array
          items: { $ref: '#/components/schemas/MemberRole' }
    EffectivePermissions:
      type: object
      required: [user_id, permissions]
      properties:
        user_id: { type: string, format: uuid }
        permissions:
          type: array
          items: { $ref: '#/components/schemas/Permission' }
...